Embedd timegm function for Solaris that doesn't have the function.
I modified three files - configure.ac, cbits/config.h.in and cbits/conv.c.

By the modification of configure.ac and cbits/config.h, HAVE_TIMEGM will be or not be defined depending on the presense of the function.
In cbits/conv.c, embedded timegm will be used if the HAVE_TIMEGM is not defined.

The embedded code was from the following URL:

 ftp://ftp.smr.ru/pub/0/FreeBSD/releases/i386/branches/-current/src/crypto/heimdal/lib/asn1/timegm.c

The license of it is the same as that of unix-time (BSD).

embedd timegm code for Solaris.

modified to use AC_CHECK_FUNCS, not the predefined __sun__

In the first modification, I used "#if defined(__sun__)" to switch, however, it was not good.
In this revision, I put AC_CHECK_FUNCS(timegm) in the configure.ac and use "HAVE_TIMEGM" in conv.c.

I also had to modify cbits/config.h.in.