Add UTC localization hack in *GMT conv, fix #36
strftime and strptime use `local time`(through mktime and tzset() calls).
We address the issue by setting the TZ environment variable to UTC
before making the call to strftime/strptime and then setting it back
to the previous value if there was one or unsetting it if there was
none.