Robustly detect _mkgmtime and _isblank_l on Windows
GHC 9.4 on Windows switches from using a GCC-based C toolchain to a Clang-based
one. This revealed that the way `unix-time`'s `configure.ac` script detects the
presence of the `_mkgmtime` and `_isblank_l` functions is fragile, as they
depend on MSVCRT-specific assumptions that do not hold when linking against
UCRT (as is the case with the Clang-based toolchain). This patch implements the
necessary tweaks to make `unix-time` use these functions in a slightly more
portable way. See
https://github.com/kazu-yamamoto/unix-time/issues/58#issuecomment-1237072121
for a more in-depth look at why this is done this way.

Fixes #58.