Windows: Link against mingwex to work around GHC#23533
GHC 9.4.5, 9.6.1, and 9.6.2 on Windows do not link against the `mingwex`
library, which `unix-time` implicitly depends on. This results in runtime
linker errors when combining `unix-time` with Template Haskell. See
https://gitlab.haskell.org/ghc/ghc/-/issues/23533 for an example of this.

We work around the issue by conditionally linking against `mingwex` on Windows.
I've guarded this behind a conditional so that it can be removed in the distant
future, should `unix-time` ever drop support for these GHC versions.