Use correct type for tv_usec when peeking UnixTime
UnixTime defines 'utMicroSeconds' as 'Int32', but 'tv_usec' (in 'struct
timeval') is of type 'suseconds_t'.

The size of 'suseconds_t' is platform specific, and it is 8 bytes long
on 64-bit platforms. This happens to give us the correct result on small
endian architectures, but results in a value of 0 for 'utMicroSeconds'
on 64-bit big-endian architectures.

In order to avoid changing the type of 'utMicroSeconds' (which will
result in an API change) we peek 'tv_usec' using the 'CSUSeconds' type
and then convert it to 'Int32' (the type of 'utMicroSeconds'). We rely
on the fact that 'tv_usec' is no bigger than '1000000' and hence we will
not overflow during the above conversion.

Refs #62